Safari-609.1.11.1/LayoutTests/accessibility/mac/misspelled-attributed-string-expected.txt

wrods is misspelled aab lotsi nowadays. euep.
This tests that misspelled words are correctly marked when asking for an attributed string.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS text.stringForRange(0, 5) is 'wrods'
PASS text.attributedStringRangeIsMisspelled(0, 5) is true
PASS text.stringForRange(6, 2) is 'is'
PASS text.attributedStringRangeIsMisspelled(6, 2) is false
PASS text.stringForRange(9, 10) is 'misspelled'
PASS text.attributedStringRangeIsMisspelled(9, 10) is false
PASS text.stringForRange(20, 3) is 'aab'
PASS text.attributedStringRangeIsMisspelled(20, 3) is true
PASS text.stringForRange(24, 5) is 'lotsi'
PASS text.attributedStringRangeIsMisspelled(24, 5) is true
PASS text.stringForRange(30, 8) is 'nowadays'
PASS text.attributedStringRangeIsMisspelled(30, 8) is false
PASS text.stringForRange(40, 5) is 'euep.'
PASS text.attributedStringRangeIsMisspelled(40, 5) is true
PASS successfullyParsed is true

TEST COMPLETE
